Issue-423: Moved IDiscoverMapper to core module
