+++ core/build.gradle
@@ -1,6 +1,8 @@
 apply from: '../buildsystem/android-module.gradle'
 
 dependencies {
+    implementation project(path: ':networkdata')
+
     implementation libs.bundles.architectureComponents
     implementation libs.bundles.internet
     implementation libs.bundles.di
